Move common fixtures to conftest to be used in other test files
